requirements.txt
exifread
aiogram
python-dotenv
pytz
fake-useragent
geopy